+++ packages/qwik/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Fix type errors after upgrading to Qwik v1.4.0
+- Upgrade Qwik and Qwik City peer dependency to v1.4.0
+
 ## v0.22.0 (December 15, 2023)
 
 - Fix type errors after upgrading to Qwik v1.3.1